requirements.txt
charset-normalizer==2.0.12
Django==3.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pandas==2.2.2
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
sqlparse==0.5.0
requests==2.26.0
